Move Upgrade operation to AdminSubscriptionOperations
The upgrade operation needs the user+app context to perform
upgrade+license transfer which is the operation most often wanted for
such upgrade operations.
