+++ src/main/java/org/springframework/social/partnercenter/api/order/subscription/SubscriptionOperations.java
@@ -7,7 +7,7 @@ import org.springframework.social.partnercenter.api.order.subscription.upgrade.U
 import org.springframework.social.partnercenter.api.order.subscription.upgrade.UpgradeResult;
 
 public interface SubscriptionOperations extends PagingResourceOperations<Subscription>{
-	ResponseEntity<Subscription> getById(String resellerCid, String id);
+	ResponseEntity<Subscription> getById(String customerTenantId, String id);
 	ResponseEntity<PartnerCenterResponse<Subscription>> getCustomersSubscriptions(String customerId);
 	ResponseEntity<PartnerCenterResponse<Subscription>> getSubscriptionsByOrderId(String customerId, String orderId);
 	ResponseEntity<PartnerCenterResponse<Subscription>> getAddOnsForBySubscriptionId(String customerId, String subscriptionId);
@@ -24,21 +24,4 @@ public interface SubscriptionOperations extends PagingResourceOperations<Subscri
 	 * @return Subscription
 	 */
 	ResponseEntity<Subscription> updateSubscriptionQuantity(String customerId, String subscriptionId, int qty);
-
-	/**
-	 * Retrieves a list of offers available fot upgrade
-	 * @param customerId Customer tenant id.
-	 * @param subscriptionId Subscription to be upgraded
-	 * @return Upgrade[]
-	 */
-	ResponseEntity<PartnerCenterResponse<Upgrade>> getAvailableUpgrades(String customerId, String subscriptionId);
-
-	/**
-	 * Upgrades a customer's subscription to a specified target subscription.
-	 * @param customerId Customer tenant id.
-	 * @param sourceSubscriptionId Subscription to be upgraded.
-	 * @param upgrade Chosen offer to upgrade to.
-	 * @return UpgradeResult
-	 */
-	ResponseEntity<UpgradeResult> upgradeSubscription(String customerId, String sourceSubscriptionId, Upgrade upgrade);
 }
