+++ src/main/java/org/springframework/social/partnercenter/api/order/subscription/impl/AdminSubscriptionTemplate.java
@@ -6,6 +6,8 @@ import org.springframework.social.partnercenter.api.PartnerCenterResponse;
 import org.springframework.social.partnercenter.api.order.subscription.AdminSubscriptionOperations;
 import org.springframework.social.partnercenter.api.order.subscription.Conversion;
 import org.springframework.social.partnercenter.api.order.subscription.ConversionResult;
+import org.springframework.social.partnercenter.api.order.subscription.upgrade.Upgrade;
+import org.springframework.social.partnercenter.api.order.subscription.upgrade.UpgradeResult;
 import org.springframework.social.partnercenter.http.client.RestResource;
 
 public class AdminSubscriptionTemplate extends SubscriptionTemplate implements AdminSubscriptionOperations {
@@ -29,5 +31,19 @@ public class AdminSubscriptionTemplate extends SubscriptionTemplate implements A
 		return restResource.request()
 				.pathSegment(customerTenantyId, SUBSCRIPTIONS, subscriptionId, CONVERSIONS)
 				.post(conversion, ConversionResult.class);
+	}
+
+	@Override
+	public ResponseEntity<PartnerCenterResponse<Upgrade>> getAvailableUpgrades(String customerId, String subscriptionId) {
+		return restResource.request()
+				.pathSegment(customerId, SUBSCRIPTIONS, subscriptionId, "upgrades")
+				.get(new ParameterizedTypeReference<PartnerCenterResponse<Upgrade>>() {});
+	}
+
+	@Override
+	public ResponseEntity<UpgradeResult> upgradeSubscription(String customerId, String sourceSubscriptionId, Upgrade upgrade) {
+		return restResource.request()
+				.pathSegment(customerId, SUBSCRIPTIONS, sourceSubscriptionId, "upgrades")
+				.post(upgrade, UpgradeResult.class);
 	}
 }
